Refactor vertex tuple
